シーン間でデータの受け渡しをできるように変更

+++ project/application/GameObjects/CommonData.h
@@ -4,6 +4,6 @@
 /// シーンをまたいで保持するデータ
 /// </summary>
 struct CommonData {
-	float killCount; // 敵の死亡数
+	int killCount; // 敵の死亡数
 	float clearTime; // クリアタイム
 };
